improvement on audio progress

+++ SessionMessagingKit/Sending & Receiving/Attachments/TSAttachmentStream.m
@@ -522,7 +522,7 @@ typedef void (^OWSLoadedThumbnailSuccess)(OWSLoadedThumbnail *loadedThumbnail);
     // This will block for at most 5s.
     // Theoritically it won't block the thread since the media file is a local file.
     VLCTime *length = [audioPlayer.media lengthWaitUntilDate:[NSDate dateWithTimeIntervalSinceNow:5.0]];
-    return length.intValue/1000;
+    return (CGFloat)length.intValue/1000;
 }
 
 - (CGFloat)audioDurationSeconds
